refactor : ChatLog Redis 설정 리팩토링
