+++ chat-service/src/main/java/webrtc/v1/MainApplication.java
@@ -5,7 +5,6 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cache.annotation.EnableCaching;
 
 @SpringBootApplication
-@EnableCaching
 public class MainApplication {
 
 	public static void main(String[] args) {
