+++ chat-service/src/main/java/webrtc/v1/chat/repository/ChatLogRedisRepository.java
@@ -6,5 +6,6 @@ public interface ChatLogRedisRepository {
 
     void save(String channelId, ChatLog chatLog);
 
+    void delete(String id);
 
 }
